clean Makefile
Using `$<` (first dependency) instead of `$^` (all) for the `%.o` rule
means the `boot.h` dependency can correctly be attached to the `.o`
(otherwise it doesn't rebuild the executable when the Lua file changes).

But actually `make` has a builtin rule for `%.o`, and so the whole thing
can go.

Also cleaned up the `clean` rule – vim `.` (repeat last edit) accident?